+++ mytorch/nn/LayerNorm.py
@@ -18,11 +18,11 @@ class LayerNorm(Module):
         self.reshaped_x = x.reshape(*self.original_shape[:self.axis],-1)
         
         if self.gamma is None:
-            self.gamma = cp.ones((*self.reshaped_x.shape[:-1],1))
-            self.beta = cp.zeros((*self.reshaped_x.shape[:-1],1))
+            self.gamma = cp.ones((*self.reshaped_x.shape[:-1],1), dtype=cp.float32)
+            self.beta = cp.zeros((*self.reshaped_x.shape[:-1],1), dtype=cp.float32)
 
-        self.mean = cp.mean(self.reshaped_x, axis=-1, keepdims=True)
-        self.std = cp.std(self.reshaped_x, axis=-1, keepdims=True)
+        self.mean = cp.mean(self.reshaped_x, axis=-1, keepdims=True, dtype=cp.float32)
+        self.std = cp.std(self.reshaped_x, axis=-1, keepdims=True, dtype=cp.float32)
         
         self.y = (x - self.mean) / (self.std + 1e-5)
         
@@ -32,11 +32,11 @@ class LayerNorm(Module):
         jacobian = self.calculate_jacobian()
         delta_error = cp.matmul(error, jacobian).reshape(self.original_shape)
         
-        delta_gamma = cp.sum(error * self.y, axis=-1)
-        delta_beta = cp.sum(error, axis=-1)
+        delta_gamma = cp.sum(error * self.y, axis=-1, dtype=cp.float32)
+        delta_beta = cp.sum(error, axis=-1, dtype=cp.float32)
         
-        self.gamma -= lr * cp.mean(delta_gamma, axis=0)
-        self.beta -= lr * cp.mean(delta_beta, axis=0)
+        self.gamma -= lr * cp.mean(delta_gamma, axis=0, dtype=cp.float32)
+        self.beta -= lr * cp.mean(delta_beta, axis=0, dtype=cp.float32)
         
         return delta_error
     
@@ -45,7 +45,7 @@ class LayerNorm(Module):
         # https://neuralthreads.medium.com/layer-normalization-and-how-to-compute-its-jacobian-for-backpropagation-55a549d5936f
         
         N = self.reshaped_x.shape[-1]
-        I = cp.eye(N)
+        I = cp.eye(N, dtype=cp.float32)
         
         first_part = (cp.tile(I, (*self.reshaped_x.shape[:-1],1,1)) - 1) / (N*self.std[...,cp.newaxis])
